# preallocate buffers when decoding/encoding way nodes