#### Here's the reason for this change :rocket:
The `idSliceToBytes` and `bytesToIDSlice` in `pbf2json.go` do not preallocate their buffers causing a performance hit to these functions.  There are few other similar places that could be updated if changes like this are okay.

---
#### Here's what actually got changed :clap:
- [x] updated `idSliceToBytes` to preallocate the bytes buffer
- [x] updated `bytesToIDSlice` to preallocate the ids buffer
- [x] cleaned `bytesToLatLon` for a minor speed improvement
- [x] cleaned up `nodeToBytes` to preallocate and reuse the buffer
- [x] wrote some benchmarks to show the speed improvement

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkBytesToLatLon-12      809           753           -6.92%
BenchmarkNodeToBytes-12        84.5          58.0          -31.36%
BenchmarkIdSliceToBytes-12     1068          232           -78.28%
BenchmarkBytesToIDSlice-12     579           252           -56.48%

benchmark                      old allocs     new allocs     delta
BenchmarkBytesToLatLon-12      8              6              -25.00%
BenchmarkNodeToBytes-12        2              2              +0.00%
BenchmarkIdSliceToBytes-12     5              1              -80.00%
BenchmarkBytesToIDSlice-12     8              1              -87.50%

benchmark                      old bytes     new bytes     delta
BenchmarkBytesToLatLon-12      448           432           -3.57%
BenchmarkNodeToBytes-12        67            19            -71.64%
BenchmarkIdSliceToBytes-12     2224          896           -59.71%
BenchmarkBytesToIDSlice-12     2040          896           -56.08%
```

---
#### Here's how others can test the changes :eyes:
There are existing tests that cover the two impacted functions.