@orangejulius 2020-12-31
Hey Paul,
Thanks for this, looks really great!

These look like extremely common operations, so I imagine there's a big speedup to general decoding of `pbf` files, do you have any rough idea of what one could expect?

P.S. I think we met once at the Mapzen office long ago, maybe after the SOTM in NYC? Hope you're doing well!