@paulmach 2021-01-01
Unfortunately I haven't seen much of a speedup when doing something like `build/pbf2json.darwin-x64 -tags="building" new-jersey-latest.osm.pbf`. I suspect a majority of the time is spent decoding the pbf. 

Ah yes, the Mapzen after party. Simpler times :)