@paulmach 2021-01-04
You're probably ahead of me on this but after playing around, if leveldb is involved it's the bottleneck, if it's not, then there is some room for improvement, but I'm not sure how common that is.