@paulmach 2021-01-07
I did a bunch of profiling over the last few days and on my osx laptop and I found that system calls to write files (and stdout) were the bottle neck. I was kind of surprised. So unless you're passing large files for a few items, there isn't much room for improvement :(