@missinglink 2021-01-07
Yeah, I suspect that using `mmap` for syscalls would yield some performance gains but probably not enough to justify the time spent on it, servers are cheap, humans are expensive ;)

Thanks for the PR, I'm on the road at the moment so I won't republish the prebuilt binaries until later.